Revert version info to build successfully
Since the github checkout action fetches only a single commit by
default, we won't have access to all the git metadata, which causes the
build to fail with an empty string defined for version name, which is
not allowed. So we're removing the git metadata from the build config
files temporarily until we figure out a more robust way of dynamically
filling in the version info both in CI and local environments.