Add a properties file for Jacoco
Jacoco crashes when the app is started up, and this crash shows up in
the logs, even though it does not affect the app. Adding this properties
file with output set to none enables Jacoco to start up without failing,
and coverage will be written as normal and transferred correctly later
by the android gradle plugin coverage implementation.